+++ src/app/(header)/_components/Content.tsx
@@ -1,34 +1,17 @@
 'use client';
 
 import { ToTopFloatingButton } from '@/components';
+import { useScroll } from '@/hooks/useScroll';
 import classNames from 'classnames';
-import { ReactNode, useRef } from 'react';
+import { ReactNode } from 'react';
 
 type ContentType = {
   children: ReactNode;
 };
 
-interface CustomDiv extends HTMLDivElement {
-  timeoutId?: NodeJS.Timeout;
-}
-
 export default function Content({ children }: ContentType) {
-  const scrollableRef = useRef<CustomDiv>(null);
+  const { scrollableRef, handleScroll } = useScroll();
 
-  const handleScroll = () => {
-    if (scrollableRef.current) {
-      scrollableRef.current.classList.add('show-scrollbar');
-      if (scrollableRef.current.timeoutId) {
-        clearTimeout(scrollableRef.current.timeoutId);
-      }
-      scrollableRef.current.timeoutId = setTimeout(function () {
-        if (scrollableRef.current) {
-          scrollableRef.current.classList.remove('show-scrollbar');
-        }
-      }, 500);
-    }
-  };
-  console.log('Content');
   return (
     <div
       ref={scrollableRef}
